web: store only job ids for precomputations
For the precomputations around jobs (`pipelineLayers`, `pipelineJobs`),
we are currently storing the entire job. This has a few drawbacks:

* Higher memory usage
* Storing duplicated data (potential to get out of sync)
* If we choose to cache the precomputations, it will take up much more
space in localStorage. The limit in localStorage is typically around
5-10MB, so if we stored the full job for precomputations, it would take
3 times as much space (and we could much more easily pass the limit)

Now, we only store the job id (team, pipeline, and job names), and look
up the job details when we need details.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>